+++ front/src/components/LoginForm.tsx
@@ -1,0 +1,39 @@
+import Container from '@mui/material/Container';
+import Stack from '@mui/material/Stack';
+
+import TextField from '@mui/material/TextField';
+import Button from '@mui/material/Button';
+
+//フォームの処理
+import { useForm } from "react-hook-form";
+
+type FormValues = {
+    userId: string,
+    password: string
+};
+
+export const LoginForm = () => {
+    const { register, handleSubmit, formState: { errors } } = useForm<FormValues>();
+    const onSubmit = handleSubmit((data) => console.log(data));
+
+    return (
+        <Container maxWidth="sm">
+            <Stack spacing={2}>
+                <br></br>
+                <h1>ログイン</h1>
+                <form onSubmit={onSubmit}>
+                <TextField {...register("userId")} id="standard-basic" label="ユーザ名" variant="standard" /><br />
+                <TextField 
+                    {...register("password")}
+                    id="standard-password-input"
+                    label="パスワード"
+                    type="password"
+                    autoComplete="current-password"
+                    variant="standard"
+                /><br />
+                <Button type='submit'>ログイン</Button>
+                </form>
+            </Stack>
+            </Container>
+    );
+}
